#49 | Keshava & Neha | refactoring
