+++ src/main/java/com/capitalone/dashboard/collector/SonarClient.java
@@ -18,7 +18,7 @@ public interface SonarClient {
      */
     void setServerCredentials(String username, String password, String token);
     List<SonarProject> getProjects(String instanceUrl);
-    CodeQuality currentCodeQuality(SonarProject project);
+    CodeQuality currentStaticCodeQuality(SonarProject project);
     CodeQuality currentSecurityCodeQuality(SonarProject project);
     JSONArray getQualityProfiles(String instanceUrl) throws ParseException;
     List<String> retrieveProfileAndProjectAssociation(String instanceUrl,String qualityProfile) throws ParseException;
