--- src/main/java/com/capitalone/dashboard/repository/SonarCollectorRepository.java
@@ -1,6 +1,0 @@
-package com.capitalone.dashboard.repository;
-
-import com.capitalone.dashboard.model.SonarCollector;
-
-public interface SonarCollectorRepository extends BaseCollectorRepository<SonarCollector> {
-}
