+++ src/main/java/com/capitalone/dashboard/repository/SonarSecurityAnalysisCollectorRepository.java
@@ -1,0 +1,6 @@
+package com.capitalone.dashboard.repository;
+
+import com.capitalone.dashboard.model.SonarSecurityAnalysisCollector;
+
+public interface SonarSecurityAnalysisCollectorRepository extends BaseCollectorRepository<SonarSecurityAnalysisCollector> {
+}
