--- src/main/java/com/capitalone/dashboard/repository/SonarSecurityCollectorRepository.java
@@ -1,6 +1,0 @@
-package com.capitalone.dashboard.repository;
-
-import com.capitalone.dashboard.model.SonarSecurityCollector;
-
-public interface SonarSecurityCollectorRepository extends BaseCollectorRepository<SonarSecurityCollector> {
-}
