+++ src/main/java/com/capitalone/dashboard/repository/SonarStaticAnalysisCollectorRepository.java
@@ -1,0 +1,6 @@
+package com.capitalone.dashboard.repository;
+
+import com.capitalone.dashboard.model.SonarStaticAnalysisCollector;
+
+public interface SonarStaticAnalysisCollectorRepository extends BaseCollectorRepository<SonarStaticAnalysisCollector> {
+}
