+++ src/test/java/com/capitalone/dashboard/collector/SonarStaticAnalysisCollectorTaskTest.java
@@ -2,10 +2,10 @@ package com.capitalone.dashboard.collector;
 
 import com.capitalone.dashboard.model.Component;
 import com.capitalone.dashboard.model.ConfigHistOperationType;
-import com.capitalone.dashboard.model.SonarCollector;
+import com.capitalone.dashboard.model.SonarStaticAnalysisCollector;
 import com.capitalone.dashboard.repository.CodeQualityRepository;
 import com.capitalone.dashboard.repository.ComponentRepository;
-import com.capitalone.dashboard.repository.SonarCollectorRepository;
+import com.capitalone.dashboard.repository.SonarStaticAnalysisCollectorRepository;
 import com.capitalone.dashboard.repository.SonarProfileRepostory;
 import com.capitalone.dashboard.repository.SonarProjectRepository;
 import org.bson.types.ObjectId;
@@ -29,10 +29,10 @@ import static org.mockito.Mockito.verifyZeroInteractions;
 import static org.mockito.Mockito.when;
 
 @RunWith(MockitoJUnitRunner.class)
-public class SonarCollectorTaskTest {
+public class SonarStaticAnalysisCollectorTaskTest {
 
-    @InjectMocks private SonarCollectorTask task;
-    @Mock private SonarCollectorRepository sonarCollectorRepository;
+    @InjectMocks private SonarStaticAnalysisCollectorTask task;
+    @Mock private SonarStaticAnalysisCollectorRepository sonarStaticAnalysisCollectorRepository;
     @Mock private SonarProjectRepository sonarProjectRepository;
     @Mock private CodeQualityRepository codeQualityRepository;
     @Mock private SonarProfileRepostory sonarProfileRepostory;
@@ -54,38 +54,38 @@ public class SonarCollectorTaskTest {
     private static final String NICENAME2 = "niceName2";
     private static final String QUALITYPROFILE = "cs-default-donotmodify-89073";
     private JSONArray qualityProfiles = new JSONArray();
-    private JSONArray profileConfigurationChanges = new JSONArray();                                                           
+    private JSONArray profileConfigurationChanges = new JSONArray();
     private JSONObject qualityProfile = new JSONObject();
     private JSONObject profileConfigurationChange = new JSONObject();
     ConfigHistOperationType operation = ConfigHistOperationType.CHANGED;
-    
+
     @Before
     public void setup() throws ParseException{
     	qualityProfile.put("key", QUALITYPROFILE);
     	qualityProfile.put("name", "Default-DoNotModify");
     	qualityProfiles.add(qualityProfile);
-    	
+
     	profileConfigurationChange.put("authorName", "foo");
     	profileConfigurationChange.put("authorLogin", "bar");
     	profileConfigurationChange.put("date", "2017-10-05T13:57:40+0000");
     	profileConfigurationChange.put("action", "DEACTIVATED");
     	profileConfigurationChanges.add(profileConfigurationChange);
-    	
+
     	Mockito.doReturn(qualityProfiles).when(defaultSonarClient).getQualityProfiles(SERVER1);
-    	
-    	Mockito.doReturn(profileConfigurationChanges).when(defaultSonarClient).getQualityProfileConfigurationChanges(SERVER1, QUALITYPROFILE);                                                 
+
+    	Mockito.doReturn(profileConfigurationChanges).when(defaultSonarClient).getQualityProfileConfigurationChanges(SERVER1, QUALITYPROFILE);
 
     	Mockito.doReturn(qualityProfiles).when(defaultSonar6Client).getQualityProfiles(SERVER1);
     	Mockito.doReturn(qualityProfiles).when(defaultSonar6Client).getQualityProfiles(SERVER2);
 
-    	Mockito.doReturn(profileConfigurationChanges).when(defaultSonar6Client).getQualityProfileConfigurationChanges(SERVER1, QUALITYPROFILE);                                                 
+    	Mockito.doReturn(profileConfigurationChanges).when(defaultSonar6Client).getQualityProfileConfigurationChanges(SERVER1, QUALITYPROFILE);
     	Mockito.doReturn(profileConfigurationChanges).when(defaultSonar6Client).getQualityProfileConfigurationChanges(SERVER2, QUALITYPROFILE);
     }
 
     @Test
     public void collectEmpty() throws Exception {
         when(dbComponentRepository.findAll()).thenReturn(components());
-        task.collect(new SonarCollector());
+        task.collect(new SonarStaticAnalysisCollector());
         verifyZeroInteractions(sonarClientSelector, codeQualityRepository);
     }
 
@@ -159,11 +159,11 @@ public class SonarCollectorTaskTest {
 
         verify(sonarClientSelector).getSonarClient(VERSION43);
         verify(sonarClientSelector).getSonarClient(VERSION54);
-        
+
         verify(defaultSonar6Client).getQualityProfiles(SERVER2);
         verify(defaultSonar6Client).retrieveProfileAndProjectAssociation(SERVER2, QUALITYPROFILE);
         verify(defaultSonar6Client).getQualityProfileConfigurationChanges(SERVER2, QUALITYPROFILE);
-        
+
     }
 
     private ArrayList<com.capitalone.dashboard.model.Component> components() {
@@ -176,12 +176,12 @@ public class SonarCollectorTaskTest {
         return cArray;
     }
 
-    private SonarCollector collectorWithOneServer() {
-        return SonarCollector.prototype(Collections.singletonList(SERVER1),Collections.singletonList(NICENAME1));
+    private SonarStaticAnalysisCollector collectorWithOneServer() {
+        return SonarStaticAnalysisCollector.prototype(Collections.singletonList(SERVER1),Collections.singletonList(NICENAME1));
     }
 
-    private SonarCollector collectorWithOnTwoServers() {
-        return SonarCollector.prototype(Arrays.asList(SERVER1, SERVER2), Arrays.asList(NICENAME1,NICENAME2));
+    private SonarStaticAnalysisCollector collectorWithOnTwoServers() {
+        return SonarStaticAnalysisCollector.prototype(Arrays.asList(SERVER1, SERVER2), Arrays.asList(NICENAME1,NICENAME2));
     }
 
 }
