+++ src/test/resources/com/capitalone/dashboard/collector/sonar6measures.json
